Fix Storybook build
