+++ eisbuk-admin/src/components/auth/Unauthorized.stories.tsx
@@ -1,25 +1,18 @@
 import React from "react";
-import Unauthorized from "./Unauthorized";
+import { ComponentStory, ComponentMeta } from "@storybook/react";
+
+import Unauthorized from "@/components/auth/Unauthorized";
 
 export default {
   title: "Unauthorized page",
   component: Unauthorized,
+} as ComponentMeta<typeof Unauthorized>;
+
+const Template: ComponentStory<typeof Unauthorized> = (args) => (
+  <Unauthorized {...args} />
+);
+
+export const Default = Template.bind({});
+Default.args = {
+  backgroundIndex: 0,
 };
-
-export const Default = (): JSX.Element => <Unauthorized backgroundIndex={0} />;
-
-export const Background1 = (): JSX.Element => (
-  <Unauthorized backgroundIndex={1} />
-);
-
-export const Background2 = (): JSX.Element => (
-  <Unauthorized backgroundIndex={2} />
-);
-
-export const Background3 = (): JSX.Element => (
-  <Unauthorized backgroundIndex={3} />
-);
-
-export const Background4 = (): JSX.Element => (
-  <Unauthorized backgroundIndex={4} />
-);
