+++ eisbuk-admin/src/components/slots/SlotForm/SlotForm.stories.tsx
@@ -1,40 +1,41 @@
 import React from "react";
-import SlotForm from "./SlotForm";
+import { ComponentMeta, ComponentStory } from "@storybook/react";
+import firebase from "firebase";
+
+import SlotForm from "@/components/slots/SlotForm";
+import { Category, Duration, SlotType } from "@/enums/firestore";
 
 export default {
   title: "SlotForm",
   component: SlotForm,
-};
+  argTypes: {
+    editSlot: { action: "udpated" },
+    createSlot: { action: "created" },
+    onClose: { action: "closed" },
+    onOpen: { action: "opened" },
+  },
+} as ComponentMeta<typeof SlotForm>;
 
 const baseProps = {
   open: true,
   isoDate: "2021-01-15",
 };
 
-export const EmptyForm = (): JSX.Element => <SlotForm {...baseProps} />;
+const Template: ComponentStory<typeof SlotForm> = (
+  args: Omit<Omit<Parameters<typeof SlotForm>[0], "open">, "isoDate">
+) => <SlotForm {...baseProps} {...args} />;
 
-// EmptyForm.argTypes = {
-//   createSlot: { action: "created" },
-//   onClose: { action: "closed" },
-// };
+export const EmptyForm = Template.bind({});
 
-const initialValues = {
-  time: "11:30",
-  categories: ["preagonismo"],
-  durations: ["60", "120"],
-  type: "ice",
-  notes: "Here are some notes\nWith two lines",
-};
-
-export const FormWithValues = (): JSX.Element => (
-  <SlotForm {...baseProps} {...initialValues} />
-);
+export const FormWithValues = Template.bind({});
 FormWithValues.args = {
-  open: true,
-  isoDate: "2021-01-15",
-  initialValues: {},
+  slotToEdit: {
+    id: "random_id",
+    date: firebase.firestore.Timestamp.now(),
+    time: "11:30",
+    categories: [Category.Preagonismo],
+    durations: [Duration["1h"], Duration["2h"]],
+    type: SlotType.Ice,
+    notes: "Here are some notes\nWith two lines" as any,
+  },
 };
-// FormWithValues.argTypes = {
-//   createSlot: { action: "created" },
-//   onClose: { action: "closed" },
-// };
