+++ eisbuk-admin/src/components/slots/SlotListByDay/SlotsDay.stories.tsx
@@ -1,15 +1,18 @@
 import React from "react";
-import { Timestamp } from "@google-cloud/firestore";
+import firebase from "firebase";
 import { v4 } from "uuid";
 
-import SlotsDay from "./SlotsDay";
+import SlotsDay from "@/components/slots/SlotListByDay/SlotsDay";
 
 import { Category, Duration, Notes, SlotType } from "@/enums/firestore";
+import { ComponentStory } from "@storybook/react";
 
 export default {
   title: "Slots Day",
   component: SlotsDay,
 };
+
+const Timestamp = firebase.firestore.Timestamp;
 
 const uuid = v4;
 
@@ -18,47 +21,51 @@ const baseProps = {
   view: "slots",
 };
 
-const multipleSlotProps = {
-  slots: {
-    foo: {
-      id: uuid(),
-      categories: [Category.Agonismo, Category.Preagonismo],
-      type: SlotType.Ice,
-      date: new Timestamp(1609513200, 0),
-      durations: [Duration["1h"]],
-      notes: Notes.Pista1,
-    },
-    bar: {
-      id: uuid(),
-      date: new Timestamp(1609495200, 0),
-      categories: [Category.Preagonismo, Category.Agonismo, Category.Corso],
-      type: SlotType.OffIceDanza,
-      durations: [Duration["1.5h"], Duration["2h"]],
-      notes: Notes.Pista2,
-    },
-    baz: {
-      id: uuid(),
-      date: new Timestamp(1609516800, 0),
-      categories: [Category.Corso],
-      type: SlotType.OffIceGym,
-      durations: Object.values(Duration),
-      notes: Notes.Pista2, // this was deleted `\nPotrebbe non svolgersi`
-    },
+const slots = {
+  foo: {
+    id: uuid(),
+    categories: [Category.Agonismo, Category.Preagonismo],
+    type: SlotType.Ice,
+    date: new Timestamp(1609513200, 0),
+    durations: [Duration["1h"]],
+    notes: Notes.Pista1,
+  },
+  bar: {
+    id: uuid(),
+    date: new Timestamp(1609495200, 0),
+    categories: [Category.Preagonismo, Category.Agonismo, Category.Corso],
+    type: SlotType.OffIceDanza,
+    durations: [Duration["1.5h"], Duration["2h"]],
+    notes: Notes.Pista2,
+  },
+  baz: {
+    id: uuid(),
+    date: new Timestamp(1609516800, 0),
+    categories: [Category.Corso],
+    type: SlotType.OffIceGym,
+    durations: Object.values(Duration),
+    notes: Notes.Pista2, // this was deleted `\nPotrebbe non svolgersi`
   },
 };
 
-export const ManySlots = (): JSX.Element => (
-  <SlotsDay {...baseProps} {...multipleSlotProps} />
-);
+const Template: ComponentStory<typeof SlotsDay> = (
+  args: Omit<Omit<Parameters<typeof SlotsDay>[0], "day">, "view">
+) => <SlotsDay {...baseProps} {...args} />;
 
-export const ManySlotsWithDelete = (): JSX.Element => (
-  <SlotsDay {...baseProps} {...multipleSlotProps} enableEdit />
-);
+export const EmptyDay = Template.bind({});
 
-export const OneSlot = (): JSX.Element => (
-  <SlotsDay {...baseProps} slots={{ slots: multipleSlotProps.slots.foo }} />
-);
+export const OneSlot = Template.bind({});
+OneSlot.args = {
+  slots: { foo: slots.foo },
+};
 
-export const EmptyDay = (): JSX.Element => (
-  <SlotsDay {...baseProps} slots={{}} />
-);
+export const ManySlots = Template.bind({});
+ManySlots.args = {
+  slots,
+};
+
+export const ManySlotsWithEdit = Template.bind({});
+ManySlotsWithEdit.args = {
+  slots,
+  enableEdit: true,
+};
